fixup: Restored branch-removed assertions
They are no longer needed, but it is probably best to remove them when
refactoring this code to replace member assignments with member
initialization code. While at it, we should consider either adding
RefCount and CbcPointer methods to guarantee object existence (similar
to std::optional::value() and std::vector::at()) OR a global function
providing that same guarantees for any pointer-like object.