configure.ac: use AC_CHECK_TOOL for ar
Use AC_CHECK_TOOL which checks the environment variable `$AR` as well, which
is useful for us in Gentoo when cross-compiling.

We could use AC_PROG_AR in newer autoconf or AM_PROG_AR in automake but
the AR_R use is hardcoded in a bunch of places so not worth it.

Bug: https://bugs.gentoo.org/911945