fix loading of dhparams for DHE with Openssl >= 3
The #elif to #if OPENSSL_MAJOR_VERSION < 3 new code block introduced
with 742236c7 does not work. Instead of trying to load parameters for
DHE, it tries to parse the input file as EC Key if any curve is given to
the tls-dh parameter, e.g.
tls-dh=prime256v1:<path-to-dhparams>/dhparams4096.pem
This causes TLS handshakes to fail if the client uses a configuration
that does not accept ECDHE.

This commit corrects this by setting the type to "DH" and to further also
correctly load dhparams into the SSL_CTX from an EVP_PKEY (in contrast
to the DH used with openssl < 3), the correct function is used.