Fix some -Werror=analyzer-use-of-uninitialized-value
Static analysis reveals a compiler bug in any/Uri.cc:

  error: use of uninitialized value '<unknown>' [CWE-457]
  [-Werror=analyzer-use-of-uninitialized-value]

The temporary object used during construction of a static
variable to represent the "static X = (A + B)" logic
may not be fully defined before the addition operator is
applied to read and change its contents.

In practice this temporary object is elided by optimization
compile stages, but may (in theory) remain for -O0 builds.