Simplified "empty haystack" and "whole haystack" conditions
The conditions are mutually exclusive, but that fact was not clear in
the official code (because that code lacked `str`).

Also adjusted "no prefix" debugs() wording to clarify that use case
description and to improve symmetry with "empty haystack" use case.

This change also avoids "insufficient input" phrasing that should be
reserved for methods throwing InsufficientInput.