fixup: Do not truncate command names
Unlike strncpy(), xstrncpy() needs output buffer size, which is not 2 in this
case. Earlier PR code introduced a new truncation bug to this already buggy
official code. Official code is buggy because it does not terminate
helper_command buffer that is later used as a c-string. This change fixes both
bugs.