fixup: AAA
https://en.cppreference.com/w/cpp/language/operator_other

"Stage 6: At least one of E2 and E3 is a pointer. In this case, ...
conversions are applied to bring them to their composite pointer type,
and the result is of that type."

https://en.cppreference.com/w/cpp/language/pointer#Composite_pointer_type

"...       if p1 is a null pointer constant, C is T2.
Otherwise, if p2 is a null pointer constant, C is T1."