Do not treat 2-byte buffer space specially
HttpStateData::maybeMakeSpaceAvailable() did not allow to read if the
buffer space was less than 2 bytes. This behavior was added in 2005
2afaba0 commit in order to compensate for delayAwareRead() inability to
read 1 byte.  Then, in 2007 478cfe9 commit the similar 2-byte limitation
was added into Icap::ModXact::virginConsume() so that both ICAP and
http.cc sharing the body pipe were consistent. Now we can remove
this special treatment for 2-byte buffers, since  in 2012 4dc2b07
commit the underlying delayAwareRead() flaw was fixed.