fixup: Discovered why the whole doneWithAdaptation() prec is needed
Its (!adaptationAccessCheckPending) part is not sufficient to know that
receivedWholeAdaptedReply implies we _stored_ all received adapted
response body bytes: We could still be waiting to exhaust the BodyPipe
buffer unless we are doneWithAdaptation()!

The precondition is probably too strong: We could store the whole
adapted response while we are still sending virgin response body bytes
to the adaptation service. However, addressing that deficiency is
probably outside this branch work scope.