Remove PR-added "keep in sync" comment
The intent behind that comment is correct, but documenting
exactly what has to be kept in sync is difficult because callers
callers may not use bp.buf().spaceSize() in their calculations
or may use other conditions. A developer reading this comment
may see the differences and wonder whether the code is broken.

We know that this code is broken, but it is broken for reasons
other than bad synchronization. We will address those problems
when fixing "emptied inBuf" bugs. At that time, we may be in a
better position to document (or automate!) these
synchronization needs.