Reuse existing printing code instead of writing custom one
XXX: This change reduces the number of random/hash bits that we extract
from the hash because our encoding changes from custom 0-9A-Z (mapping 4
hash bytes to 36^4 output variations) to 0-9A-F hex (mapping 2 hash
bytes to all of their 2^16 variations). AFAICT, that is significant
reduction: Custom code was reflecting more than 20 random bits (because
2^20 < 36^4 < 2^21) while this code reflects just 16, a 20% reduction!

This change slows PidFilenameHash() down, but the function is not on a
performance-sensitive path, so we should favor code reuse. In the
foreseeable future, memcpy() will be gone, and this code will become
very simple as we would be able to reuse hash printing API that is also
needed for printing Store keys.