Fix authentication module tests
Squid was using the existence of 'auth_param ... program'
setting to determine whether to initialize and/or touch
HTTP authentication headers for the scheme.

That parameter alone is not sufficient to determine that
the scheme is properly configured, and initialized. All
schemes require additional configuration and/or state.

Use the SchemeConfig::configured() method consistently
to detemine whether the scheme has been fully configured.

Use the SchemeConfig::active() method consistently to
determine whether the scheme has been initialized and
is ready to handle traffic before allowing HTTP headers
to be produced that use it.