Simplify SBuf-to-String conversion
There is still a lot of code that uses deprecated String class, and an
increasing amount of code that has to covert modern SBufs to Strings.

SBuf handles NUL bytes better than String. With respect to handling NUL
bytes, the new conversion methods are as (un)safe as existing size-based
ones (that callers tend to use prior to these changes anyway).