Bug 5499: src_as/dst_as do not match if WHOIS requires cache_peer
Squid ACL initialization code calls asnCacheStart() and tries
to connect to an ASN server. If the configuration requires a cache_peer,
the connection fails because none of cache_peers are available an that
point:

    WARNING: AS ... whois request failed

Since ASN-based ACLs are essentially unused and fixing it requires
much work, it was decided to to drop ASN support altogether. In rare
occasions where it is still needed, an external ACL helper can be
written to provide the same functionality.