CI failure after removing CONTRIBUTOR entry
source-maintenance.sh recognizes a magic phrase in the commit message
and treats that commit as authoritative for CONTRIBUTORS file. However
when dealing with PRs, the merge commit that is built has a
GitHub-generated commit message which obviously lacks the magic
phrase.

The problem happens when, for example, a user creates a PR and wants to
change their CONTRIBUTORS entry (e.g., from U@foo to U@bar). In this
case GitHub actions start source-maintenance.sh for the merge commit
which adds the old U@foo to CONTRIBUTORS (because the U@foo is the merge
commit author). The build is aborted due to the "modified sources"
error. To avoid this problem, a PR author now can mark the
entire PR as authoritative for CONTRIBUTORS by adding the magic phrase
to the PR description.