Fix authentication module tests (#2098)
Squid was using the existence of an 'auth_param ... program'
setting to determine whether to initialize the scheme module for
use and advertise HTTP authentication headers for the scheme.

That parameter alone is not sufficient to determine that the
scheme is properly configured, and initialized. All schemes
require additional configuration and/or state.

Use the existing SchemeConfig::configured() method consistently
to determine whether the scheme has been fully configured.

Use the existing SchemeConfig::active() method consistently to
determine whether the scheme has been initialized and is ready
to handle traffic before allowing HTTP headers to be produced.