Remove unnecessary char buffer manipulation in snmp parsing
The decoded Community is not string-like, and as such, deos not need
to be null-terminated when being decoded. It needs to be
null-terminated when it is parsed however, so this is done higher in
the call-stack, after the decoding (if successful).

'bufp' is allocated with 128+1 bytes, allowing for a final null-byte.