Register icp listener after connection setup.
Among other things, this patch ensures that derefs of
`icpOutgoingConn->local` do not result in null-pointer deref. For
example, when Config.Addrs.udp_outgoing.isNoAddr() is true,
icpOutgoingConn is not set in icpOpenPorts(). It is later assigned
in icpIncomingConnectionOpened() after the read handler is installed.
That short window allows icpHandleUdp() to run with
icpOutgoingConn == nullptr.