fixup: Do not add unnecessary conditional compilation guards
In modern Squid, falsy USE_IPV6 does not mean "do not compile IPv6
code". It means, roughly speaking, "do not listen on and connect to IPv6
addresses".

Fortunately, the official `if`-statement in question is already written
to match the above meaning: Icmp6 class and the rest of this code
already compiles when USE_IPV6 is falsy, with the exception of
icmp6.Open() call. This `if`-statement icmp6_worker condition already
protects us from calling icmp6.Recv() when USE_IPV6 is falsy.