fixup: Undo move towards "callers may supply any payload" design
That design requires more code (and more decisions) to support properly.
We can and, hence, should keep it simple instead. Others may not share
this opinion, but I disagree with PR-proposed TODO (removed by this
commit). This PR is unlikely to be the right place to attempt to form
consensus, so we keep the official assertion (for now).