Use default initializers for ACLFilledChecklist
This change is not necessary to build this branch in my environment,
but, without it, some ACLFilledChecklist data members will be left
uninitialized by that class constructor STUBs, and some compilers or
code analysis tools may notice that and complain about it.

Alternatively, we can duplicate that non-trivial inialization code
yet again.