fixup: Make OpenSSL debugs() more failed-call-specific
There are several OpenSSL error-reporting layers/APIs, and the earlier
implementation was meant to emphasize which layer was being reported,
but developers will be able to look those details up in Squid sources.
Meanwhile, reporting the failed call makes this code more similar to
other error-reporting code and allows us to reuse the same approach for
other calls that may fail in these functions.